% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.missing&correlation.R
\name{check_correlation}
\alias{check_correlation}
\title{Check Correlation Between Variables}
\usage{
check_correlation(data, continuous_vars, dir_output)
}
\arguments{
\item{data}{A data frame containing the variables to analyze}

\item{continuous_vars}{Vector of column names for continuous variables}

\item{dir_output}{Directory where correlation heatmaps will be saved}
}
\value{
A vector of variable names that have strong correlations (abs(cor) > 0.5)
}
\description{
Calculates and visualizes correlations between variables in the dataset,
using different correlation methods depending on variable types.
}
\details{
This function performs the following:
\itemize{
\item Calculates pairwise correlations for continuous variables
\item Calculates correlations between all variables using appropriate methods:
\itemize{
\item Pearson for numeric-numeric pairs
\item Polychoric for categorical-categorical pairs
\item Polyserial for numeric-categorical pairs
}
\item Creates heatmaps for visualizing correlations
\item Tracks sample sizes used for each correlation pair
\item Identifies variables with high correlation (|r| > 0.5)
}
}
\examples{
\dontrun{
# Create an output directory
dir.create("output_dir")

# Get correlated variables
cor_vars <- check_correlation(
  data = your_data,
  continuous_vars = c("age", "weight", "height"),
  dir_output = "output_dir/"
)
}

}
