% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.imputation.R
\name{MI_func}
\alias{MI_func}
\title{Multiple Imputation Function}
\usage{
MI_func(
  data_selected,
  dir_output,
  info_vars = NULL,
  list_noNA = c(0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, 0.45, 0.5),
  ntree = 100,
  maxiter = 10,
  k_values = c(3, 5, 7),
  mice_m = 5,
  mice_maxit = 10,
  micer_num_trees = 100,
  seed = 123,
  niter = 10,
  selected_features = NULL,
  methods = c("rf", "knn", "mice", "mice_rf")
)
}
\arguments{
\item{data_selected}{A data frame containing the data to be imputed.}

\item{dir_output}{Directory where output files will be saved.}

\item{info_vars}{Optional data frame with identifier variables.}

\item{list_noNA}{A vector of missing rates to simulate.}

\item{ntree}{Number of trees for random forest methods.}

\item{maxiter}{Maximum number of iterations.}

\item{k_values}{A vector of K values to test for KNN imputation.}

\item{mice_m}{Number of multiple imputations for MICE.}

\item{mice_maxit}{Maximum iterations for MICE.}

\item{micer_num_trees}{Number of trees for miceRanger.}

\item{seed}{Random seed for reproducibility.}

\item{niter}{Number of iterations for each missing rate simulation.}

\item{selected_features}{Optional vector of feature names to use for imputation.}

\item{methods}{Methods used to simulate imputation.}
}
\value{
A list containing:
\itemize{
\item imputed_data: The final imputed dataset.
\item all_imputations: A list of all imputed datasets (if using multiple imputation).
\item performance: Performance metrics for each method.
\item combined_plot: Combined line plots of model performance.
\item best_model: Information about the best-performing model.
\item best_model_text: Name of best-performing model.
\item missing_summary: Missing data counts comparison before and after imputation.
\item summary_stats: Statistical comparison before and after imputation.
\item final_boxplot: Box plots of continuous variables before and after imputation.
\item final_density_plot: Density plots of continuous variables before and after imputation.
\item final_barplot: Bar plots of categorical variables before and after imputation.
\item output_file_name: The name of the output file where imputed data is saved.
\item methods_used: The name of used methods.
}
\description{
This function performs multiple imputation using various methods
(Random Forest, KNN, MICE, and miceRanger) and compares their performance.
}
\examples{
\dontrun{
result <- MI_func(
  data_selected = your_data,
  dir_output = "output/",
  list_noNA = c(0.1, 0.2, 0.3),
  ntree = 100,
  maxiter = 10,
  k_values = c(3, 5, 7),
  mice_m = 5,
  mice_maxit = 10,
  micer_num_trees = 100,
  seed = 123,
  niter = 5,
  selected_features = selected_features,
  methods = c("rf", "knn", "mice", "mice_rf")
)
}

}
